requirements.txt
python-telegram-bot>=20.7
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml>=5.1.0
APScheduler==3.10.4
cloudscraper==1.2.71
